feat: 🎸 adding events in monwatch
adding events to emit in some cases

BREAKING CHANGE: yes, monwatch now exports only class instance, but you can instantiate
how many instances you need, with different handlers
